Small amout of reorganization and renaming.
