+++ up-streamer/src/lib.rs
@@ -17,7 +17,6 @@
 //! transports.
 
 mod route;
-mod sender_wrapper;
 mod ustreamer;
 mod utransport_builder;
 mod utransport_router;
