--- up-streamer/src/sender_wrapper.rs
@@ -1,62 +1,0 @@
-/********************************************************************************
- * Copyright (c) 2024 Contributors to the Eclipse Foundation
- *
- * See the NOTICE file(s) distributed with this work for additional
- * information regarding copyright ownership.
- *
- * This program and the accompanying materials are made available under the
- * terms of the Apache License Version 2.0 which is available at
- * https://www.apache.org/licenses/LICENSE-2.0
- *
- * SPDX-License-Identifier: Apache-2.0
- ********************************************************************************/
-
-use async_std::channel::Sender;
-use std::fmt;
-use std::fmt::Debug;
-use std::hash::{Hash, Hasher};
-use std::ops::Deref;
-use std::sync::Arc;
-use up_rust::{UUIDBuilder, UUID};
-
-#[derive(Clone)]
-pub(crate) struct SenderWrapper<T> {
-    pub(crate) id: UUID,
-    sender: Arc<Sender<T>>,
-}
-
-impl<T> Debug for SenderWrapper<T> {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "SenderWrapper {{ id: {} }}", self.id)
-    }
-}
-
-impl<T> SenderWrapper<T> {
-    pub fn new(sender: Sender<T>) -> Self {
-        let id = UUIDBuilder::new().build();
-        let sender = Arc::new(sender);
-        Self { id, sender }
-    }
-}
-
-impl<T> Deref for SenderWrapper<T> {
-    type Target = Sender<T>;
-
-    fn deref(&self) -> &Self::Target {
-        &self.sender
-    }
-}
-
-impl<T> Hash for SenderWrapper<T> {
-    fn hash<H: Hasher>(&self, state: &mut H) {
-        self.id.hash(state);
-    }
-}
-
-impl<T> PartialEq for SenderWrapper<T> {
-    fn eq(&self, other: &Self) -> bool {
-        self.id == other.id
-    }
-}
-
-impl<T> Eq for SenderWrapper<T> {}
